Require that the FTS3 config is readable.
If the file `/etc/fts3/fts3config` exists - but isn't readable by
the WSGI process - then an exception is thrown by the ConfigParser.

This ensures that the file both exists and is readable.

Without it, the fts web monitoring application fails by default if
the `fts-server` RPM is also installed on the host.