FTS-1736: Replace job progress bar with a progress bar per file transfer
The previous progress bar showed the relationship between the total number of bytes transferred against the sum of the sizes of all files in a Job. Since FTS does not know the file sizes before starting a transfer the target of the progress bar was constantly changing making it confusing.

Since FTS does not know all the file sizes beforehand it only makes sense to show the progress bar on the file level.

The job summary page was changed to show the total number of bytes transferred and the total number of files already transferred. The Job progress bar was removed and an individual progress by per file was added.